perf: when choosing 'Toggle react profiling', also record mergeProps 'wasted' time
This option now also counts all time that is spent in all component mergeProps functions that results in no properties actually changing (and therefore no render). This is an important metric to improve since this is blocking the renderer window UI thread, and the time is likely spent re-computing functions even though the state input to those functions has not changed.